+++ R/gtsummary-package.R
@@ -2,6 +2,7 @@
 #' @import rlang
 #' @importFrom dplyr across
 #' @importFrom glue glue glue_data
+#' @importFrom utils head tail
 "_PACKAGE"
 
 ## usethis namespace: start