+++ R/modify_abbreviation.R
@@ -1,7 +1,7 @@
 #' Modify Abbreviations
 #'
 #' All abbreviations will be coalesced when printing the final table into
-#' a single source note.
+#' a single specialized source note.
 #'
 #' @inheritParams modify_footnote2
 #' @param abbreviation (`string`)\cr
@@ -10,6 +10,7 @@
 #'
 #' @return Updated gtsummary object
 #' @name modify_abbreviation
+#' @seealso [Footnotes vs Source Notes vs Abbreviations](https://www.danieldsjoberg.com/gtsummary/articles/modify-functions.html)
 #'
 #' @examplesIf (identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true")) && gtsummary:::is_pkg_installed(c("broom", "broom.helpers"))
 #' # Example 1 ----------------------------------
@@ -30,7 +31,7 @@ NULL
 
 #' @export
 #' @rdname modify_abbreviation
-modify_abbreviation <- function(x, abbreviation, text_interpret = c("md", "html")) {
+modify_abbreviation <- function(x, abbreviation, text_interpret = c("md", "html", "none")) {
   set_cli_abort_call()
   updated_call_list <- c(x$call_list, list(modify_footnote_body = match.call()))
 
@@ -95,7 +96,7 @@ remove_abbreviation <- function(x, abbreviation = NULL) {
       dplyr::tibble(
         column = column,
         abbreviation = abbreviation,
-        text_interpret = paste0("gt::", text_interpret)
+        text_interpret = .interpret_fun(text_interpret)
       )
     )
   x